# documentation/api.yaml
...
openapi: 3.0.0
info:
  title: Voice network API
  version: 1.0.0
  license:
    name: ISC
  description: ""
paths:
...
  "/comments/{id}":
    post:
      description: Post a comment
      summary: Post a new comment
      parameters:
        - in: path
          name: id
          description: ID of the post to comment on
          required: true
      requestBody:
        content:
          application/json:
            schema:
              properties:
                comment:
                  type: string
                  example: this a comment
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  repliesCount:
                    type: number
                    example: 0
                  _id:
                    type: string
                    example: 60136bb4a4723b3242f748e0
                  createdAt:
                    type: string
                    example: "2021-01-29T01:58:12.381Z"
                  updatedAt:
                    type: string
                    example: "2021-01-29T01:58:12.381Z"
                  body:
                    type: string
                    example: maidul98
                  post:
                    type: string
                    example: 600fae1ed984192830e4474d
                  user:
                    required:
                      - verified
                      - img_location
                      - _id
                      - username
                      - createdAt
                      - __v
                    properties:
                      verified:
                        type: boolean
                        example: false
                      img_location:
                        type: string
                        example: https://audio-social-network-profiles.s3.us-east-2.amazonaws.com/1611807455351.jpg
                      _id:
                        type: string
                        example: 600fadc1d984192830e4473a
                      username:
                        type: string
                        example: maidul98
                      createdAt:
                        type: string
                        example: "2021-01-26T05:50:57.746Z"
                      __v:
                        type: number
                        example: 0
                    type: object
        "500":
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Something went wrong, please try again
      tags:
        - Comment
      security:
        - authorization: []
  "/comments/new/{post_id}":
    get:
      summary: "Get a list of comments of a post ordered by time: newest first"
      parameters:
        - in: path
          name: post_id
          description: ID of the post
          required: true
        - in: query
          name: skip
          description:
            How many to skip before returning the next batch. You get 10
            comments with skip = 0, to get the next 10, set skip=10...
      responses:
        "200":
          description: A list of comments
        "500":
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Something went wrong, please try again
      tags:
        - Comment
      security:
        - authorization: []
  "/posts/delete/{id}":
    delete:
      description: Soft deletes the users post
      parameters:
        - in: path
          name: id
          description: ID of the post to edit
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Your post has been deleted
        "500":
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Couldn't delete your post, please try again
      tags:
        - Post
      security:
        - authorization: []
  "/posts/edit/{id}":
    put:
      parameters:
        - in: path
          name: id
          description: ID of the post to edit
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                new_caption:
                  type: string
                  example: this is a new caption
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Your post has been updated
        "400":
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Post not found
        "500":
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Couldn't update your post, please try again
      tags:
        - Post
      security:
        - authorization: []
...
  "/posts/single/{id}":
    get:
      summary: "/posts/single/{id}"
      parameters:
        - name: id
          in: path
          required: true
          description: the id of the post to be returned
          schema:
            type: string
      responses: {}
      tags:
        - Post
  "/posts/trending-posts":
    get:
      summary: "/posts/trending-posts"
      parameters:
        - in: query
          name: skip
          description:
            From where you want the next batch. The API gives you the first
            10, if you want the next 10 then you must skip 10 to get 11th post and onward
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    comments_count:
                      type: number
                    _id:
                      type: string
                    caption:
                      type: string
                    user:
                      properties:
                        verified:
                          type: boolean
                        suspended:
                          type: boolean
                        _id:
                          type: string
                        username:
                          type: string
                        createdAt:
                          type: string
                        __v:
                          type: number
                      nullable: true
                      type: object
                    audio_key:
                      type: string
                    createdAt:
                      type: string
                    __v:
                      type: number
                    votes:
                      type: object
                      properties:
                        upvoters:
                          type: array
                          items:
                            type: string
                        downvoters:
                          type: array
                          items: {}
                        voteCounts:
                          type: number
                        _id:
                          type: string
                        post:
                          type: string
                        createdAt:
                          type: string
                        __v:
                          type: number
                  required:
                    - comments_count
                    - _id
                    - caption
                    - user
                    - audio_key
                    - createdAt
                    - __v
                    - votes
        "304":
          description: ""
      tags:
        - Newsfeed
  "/posts/new":
    post:
      summary: "/posts/new"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio_file:
                  description: The audio file to upload. Must be in mp3 format.
                  type: string
                  format: binary
                caption:
                  description: Caption of the post
                  type: string
              required:
                - audio_file
                - caption
        required: true
        x-s2o-overloaded: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments_count:
                    type: number
                    example: 0
                  _id:
                    type: string
                    example: 600d10bfea27206a4550fac8
                  caption:
                    type: string
                    example: rant about ECE
                  user:
                    type: string
                    example: 6009a830d3d378157dcd4320
                  audio_key:
                    type: string
                    example: 1611468989411.mp3
                  createdAt:
                    type: string
                    example: "2021-01-24T06:16:31.019Z"
                  __v:
                    type: number
                    example: 0
                  votes:
                    type: object
                    properties:
                      upvoters:
                        type: array
                        items:
                          type: string
                        example:
                          - 6009a830d3d378157dcd4320
                      downvoters:
                        type: array
                        items: {}
                        example:
                          -
                      voteCounts:
                        type: number
                        example: 1
                      _id:
                        type: string
                        example: 600d10bfea27206a4550fac9
                      post:
                        type: string
                        example: 600d10bfea27206a4550fac8
                      createdAt:
                        type: string
                        example: "2021-01-24T06:16:31.060Z"
                      __v:
                        type: number
                        example: 0
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Only .mp3 file is allowed
      tags:
        - Post
      security:
        - authorization: []
      x-s2o-warning: Operation post:/posts/new has multiple requestBodies
  "/posts/audio/{audio_key}":
    get:
      summary: "/posts/audio/{audio_key}"
      parameters:
        - name: audio_key
          in: path
          required: true
          description: The audio_key of the post
          schema:
            type: string
      responses:
        "200":
          description: OK
          type: object
          content:
            "*/*":
              schema:
                properties:
                  audio_location:
                    type: string
                    example: https://audio-social-network.s3.us-east-2.amazonaws.com/1611466568998.mp3?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAI7KO3LWVULGZQIFA%2F20210125%2Fus-east-2%2Fs3%2Faws4_request&X-Amz-Date=20210125T032105Z&X-Amz-Expires=300&X-Amz-Signature=ca9ba5052332040af6475866478ebce7a3e2dfc9539898f91d94413ba5cb93d0&X-Amz-SignedHeaders=host
      tags:
        - Post
      description: Will return a signed URL which will expire after some time
      security:
        - authorization: []
...
servers:
  - url: https://powerful-retreat-02063.herokuapp.com
  - url: http://localhost:3000
components:
  securitySchemes:
    authorization:
      name: authorization
      in: header
      type: apiKey
...